Move new special stmts
